Revert "test: Programa with auth"
This reverts commit bdab3f176a0dfa80cb2b572fe4f518ed037536d0.
